created a main menu to add games via shortcuts and see rating changes without having to run the script every time
